+++ src/Sass/index.scss
@@ -10,6 +10,7 @@
 @import '../components/CreateUser/CreateUser';
 @import '../components/Form/Form';
 @import '../components/DynamicForm/DynamicForm';
+@import '../components/Loading/Loading';
 
 
 *, *:before, *:after {
