+++ src/components/DynamicForm/index.js
@@ -3,22 +3,23 @@ import { connect } from 'react-redux';
 import { postUser } from '../../thunks/postUser';
 import { postPart } from '../../thunks/postPart';
 import { Formik, Field } from 'formik';
+import { formValues, formConfig, selectSchema } from '../../Utility/Config/FormConfig';
 import { withRouter } from 'react-router';
 import * as Yup from 'yup';
 
-const SignupSchema = Yup.object().shape({
-  notes: Yup.string()
-    .min(1, 'Too Short!')
-    .required('Required'),
-});
+// const SignupSchema = Yup.object().shape({
+//   notes: Yup.string()
+//     .min(1, 'Too Short!')
+//     .required('Required'),
+// });
 
-const partSchema = Yup.object().shape({
-  name: Yup.string()
-    .min(1, 'Too Short!')
-    .required('Required'),
-  inventory: Yup.string()
-    .required('Required')
-});
+// const partSchema = Yup.object().shape({
+//   name: Yup.string()
+//     .min(1, 'Too Short!')
+//     .required('Required'),
+//   inventory: Yup.string()
+//     .required('Required')
+// });
 
 
 
@@ -31,20 +32,14 @@ const CreateTicket = ({formConfig, postTicket, location, history}) => (
         {
           location.formProp && formValues[location.formProp]
         }
-      onSubmit={ async (values, actions) => {
+      onSubmit={(values, actions) => {
        
-        const newUser = {...values, role: 'admin'};
-
-        console.log(values);
-        console.log(history)
-        
-        // const result =  await postUser(newUser);
 
         const formType = location.formProp;
         
         switch (formType) {
         case 'tickets':
-          await postTicket(
+          postTicket(
             {
               ...values, table_key: 1,
               table_name: 'Resources',
@@ -53,7 +48,7 @@ const CreateTicket = ({formConfig, postTicket, location, history}) => (
           history.push('/resources');
           break;
         case 'parts':
-          await postPart({
+          postPart({
             ...values, 
           }, location.itemId);
           break;
@@ -62,15 +57,14 @@ const CreateTicket = ({formConfig, postTicket, location, history}) => (
         actions.setSubmitting(false);
         
       }}
-
-      validationSchema={partSchema}
+      // validationSchema={}
       render={(props) => {
         const BASE_PROPS = {
           onChange: props.handleChange,
           onBlur: props.handleBlur
         };
         const inputNodes = location.formProp && formConfig[location.formProp].map(({html_tag, type, name, placeholder, value, label}, inputIx) => {
-          console.log(location)
+        
           return (  
             <div key={inputIx} id='radio-wrapper'>
               { type === 'radio' && 
@@ -131,6 +125,9 @@ const CreateTicket = ({formConfig, postTicket, location, history}) => (
   </section>
 );
     
+CreateTicket.defaultProps = {
+  formConfig
+};
 
 export const mapStateToProps = state => ({
   user: state.user
@@ -143,122 +140,3 @@ export const mapDispatchToProps = dispatch => ({
 export default withRouter(
   connect(null, mapDispatchToProps)(CreateTicket)
 );
-
-
-const formValues = {
-  tickets: {
-    notes: '', 
-    priority: ''
-  },
-  parts: {
-    name: '',
-    inventory: ''
-  },
-  resources: {
-    cost: '',
-    name: ''
-  },
-  resourceTypes: {
-    category: '',
-    company: '',
-    contact_number: 0,
-    contract_name: ''
-  }
-
-};
-
-
-CreateTicket.defaultProps = {
-  formConfig: {
-    tickets: [
-      {
-        html_tag: 'input',
-        type: 'radio',
-        name: 'priority',
-        value: 'low',
-        label: 'low'
-      },
-      {
-        html_tag: 'input',
-        type: 'radio',
-        name: 'priority',
-        value: 'medium',
-        label: 'medium'
-      },
-      {
-        html_tag: 'input',
-        type: 'radio',
-        name: 'priority',
-        value: 'high',
-        label: 'high'
-      },
-      {
-        html_tag: 'input',
-        type: 'radio',
-        name: 'priority',
-        value: 'urgent',
-        label: 'urgent'
-      },
-      {
-        html_tag: 'input',
-        type: 'radio',
-        name: 'priority',
-        value: 'safety',
-        label: 'safety'
-      },
-      {
-        html_tag: 'textarea',
-        name: 'notes',
-        placeholder: 'Enter Ticket Notes'
-      }
-    ],
-    parts: [
-      {
-        html_tag: 'input',
-        name: 'name',
-        type: 'text',
-        placeholder: 'Enter Part Name'
-      },
-      {
-        html_tag: 'input',
-        type: 'number',
-        name: 'inventory',
-        placeholder: 'Enter Number'
-      }
-    ],
-    resources: [
-      {
-        html_tag: 'input',
-        type: 'text',
-        name: 'name',
-        placeholder: 'Enter Name'
-      },
-      {
-        html_tag: 'input',
-        type: 'number',
-        name: 'cost',
-        placeholder: 'Enter Cost'
-      }
-    ],
-    resourceType: [
-      {
-        html_tag: 'input',
-        type: 'text',
-        name: 'company',
-        place: 'Enter Company'
-      },
-      {
-        html_tag: 'input',
-        type: 'text',
-        name: 'category',
-        placeholder: 'Enter Category'
-      },
-      {
-        html_tag: 'input',
-        type: 'number',
-        name: 'contract number',
-        placeholder: 'Enter Number'
-      }
-    ]
-  }
-};
