+++ src/components/Loading/Loading.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+const Loading = () => {
+
+  return (
+
+    <section className='loading-container'>
+      <h4>Loading&middot;&middot;&middot;</h4>
+      <div className='loading-bar'></div>
+      <div className='loading-bar'></div>
+      <div className='loading-bar'></div>
+    </section>
+    
+  );
+};
+
+export default Loading;
+
