+++ src/thunks/fetchCollection/index.js
@@ -3,32 +3,35 @@ import {
   setTickets,
   setResources,
   setLoading,
-  setError
+  setError,
+  setResourceTypes
 } from '../../actions';
 import startCase from 'lodash/startCase';
 
 const ACTION_MAP = {
   setParts,
   setTickets,
-  setResources
-}
+  setResources,
+  setResourceTypes
+};
 
 export const fetchCollection = (pathname) => {
+
   return async (dispatch) => {
     const url = `${process.env.REACT_APP_BASEURL}/api/v1${pathname}`;
     const dataKey = url.split('/').slice(-1)[0];
     const actionName = `set${startCase(dataKey)}`;
-
+console.log('url', url)
     try {
       dispatch(setLoading(true));
       const response = await fetch(url);
-
+console.log('response', response)
       if (!response.ok) {
         dispatch(setError(response.statusText));
       }
 
       const data = await response.json();
-      
+      console.log('data', data)
       dispatch(setLoading(false));
       dispatch(ACTION_MAP[actionName] && ACTION_MAP[actionName](data));
     } catch (error) {
